// src/shared/proto/rev289.gen.ts
export interface PacketDef {
    readonly name: string;
    /** Byte length; -1 means a u8 length prefix, -2 means u16. */
    readonly length: number;
}

export const REVISION = 289;

/** Indexed by opcode. null means the server never sends it at this revision. */
export const SERVER_PROT: ReadonlyArray<PacketDef | null> = [
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    { name: 'TUT_OPEN', length: 2 },
    { name: 'CHAT_FILTER_SETTINGS', length: 3 },
    null,
    null,
    null,
    null,
    { name: 'IF_SETOBJECT', length: 6 },
    null,
    null,
    { name: 'SET_PLAYER_OP', length: -1 },
    null,
    { name: 'IF_CLOSE', length: 0 },
    null,
    null,
    null,
    null,
    { name: 'UPDATE_INV_STOP_TRANSMIT', length: 2 },
    { name: 'MIDI_JINGLE', length: 4 },
    { name: 'IF_SETPLAYERHEAD', length: 2 },
    null,
    null,
    null,
    null,
    { name: 'P_COUNTDIALOG', length: 0 },
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    { name: 'UPDATE_RUNWEIGHT', length: 2 },
    { name: 'UPDATE_IGNORELIST', length: -2 },
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    { name: 'IF_OPENMAIN_SIDE', length: 4 },
    null,
    null,
    null,
    { name: 'IF_SETTEXT', length: -2 },
    { name: 'OBJ_ADD', length: 5 },
    null,
    null,
    { name: 'IF_SETTAB', length: 3 },
    null,
    { name: 'NPC_INFO', length: -2 },
    null,
    null,
    null,
    null,
    null,
    { name: 'OBJ_DEL', length: 3 },
    null,
    { name: 'CAM_MOVETO', length: 6 },
    null,
    { name: 'VARP_SMALL', length: 3 },
    { name: 'UPDATE_INV_PARTIAL', length: -2 },
    null,
    null,
    { name: 'IF_SETPOSITION', length: 6 },
    null,
    { name: 'IF_OPENCHAT', length: 2 },
    { name: 'CAM_LOOKAT', length: 6 },
    { name: 'LOC_MERGE', length: 14 },
    null,
    null,
    null,
    { name: 'MAP_PROJANIM', length: 15 },
    null,
    null,
    { name: 'LOC_ADD_CHANGE', length: 4 },
    null,
    null,
    null,
    null,
    null,
    null,
    { name: 'VARP_LARGE', length: 6 },
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    { name: 'LOC_ANIM', length: 4 },
    { name: 'UPDATE_INV_FULL', length: -2 },
    null,
    null,
    null,
    null,
    { name: 'UPDATE_ZONE_PARTIAL_ENCLOSED', length: -2 },
    null,
    null,
    { name: 'HINT_ARROW', length: 6 },
    null,
    { name: 'OBJ_COUNT', length: 7 },
    null,
    { name: 'IF_OPENMAIN', length: 2 },
    { name: 'UPDATE_PID', length: 3 },
    { name: 'LOGOUT', length: 0 },
    null,
    null,
    null,
    null,
    null,
    { name: 'IF_OPENOVERLAY', length: 2 },
    null,
    null,
    null,
    null,
    null,
    { name: 'CAM_RESET', length: 0 },
    null,
    null,
    { name: 'MINIMAP_TOGGLE', length: 1 },
    null,
    { name: 'IF_SETHIDE', length: 3 },
    null,
    null,
    null,
    null,
    null,
    { name: 'UPDATE_ZONE_FULL_FOLLOWS', length: 2 },
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    { name: 'UPDATE_STAT', length: 6 },
    { name: 'UPDATE_ZONE_PARTIAL_FOLLOWS', length: 2 },
    null,
    null,
    null,
    null,
    { name: 'IF_SETCOLOUR', length: 4 },
    null,
    null,
    null,
    { name: 'UNSET_MAP_FLAG', length: 0 },
    null,
    null,
    null,
    { name: 'UPDATE_FRIENDLIST', length: 9 },
    null,
    null,
    null,
    { name: 'RESET_CLIENT_VARCACHE', length: 0 },
    null,
    null,
    null,
    { name: 'OBJ_REVEAL', length: 7 },
    { name: 'SYNTH_SOUND', length: 5 },
    null,
    null,
    null,
    { name: 'TUT_FLASH', length: 1 },
    null,
    null,
    { name: 'IF_SETSCROLLPOS', length: 4 },
    null,
    null,
    { name: 'MIDI_SONG', length: 2 },
    { name: 'PLAYER_INFO', length: -2 },
    { name: 'IF_SETTAB_ACTIVE', length: 1 },
    null,
    null,
    null,
    null,
    { name: 'LOC_DEL', length: 2 },
    { name: 'UPDATE_RUNENERGY', length: 1 },
    { name: 'MESSAGE_GAME', length: -1 },
    null,
    null,
    null,
    null,
    { name: 'RESET_ANIMS', length: 0 },
    null,
    null,
    { name: 'UPDATE_REBOOT_TIMER', length: 2 },
    null,
    null,
    null,
    { name: 'CAM_SHAKE', length: 4 },
    null,
    null,
    { name: 'IF_SETANIM', length: 4 },
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    { name: 'REBUILD_NORMAL', length: 4 },
    null,
    null,
    { name: 'IF_SETMODEL', length: 4 },
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    { name: 'MAP_ANIM', length: 6 },
    null,
    { name: 'FRIENDLIST_LOADED', length: 1 },
    null,
    null,
    null,
    null,
    null,
    null,
    null,
    { name: 'MESSAGE_PRIVATE', length: -1 },
    { name: 'IF_SETNPCHEAD', length: 4 },
    null,
    null,
    { name: 'SET_MULTIWAY', length: 1 },
    null,
    null,
    null,
    null,
    { name: 'IF_OPENSIDE', length: 2 },
    { name: 'LAST_LOGIN_INFO', length: 10 },
    null,
    null,
];